- Checking correct answers from users.

+++ highScore/models.py
@@ -5,7 +5,7 @@ from django.conf import settings
 # Create your models here.
 class HighScore(models.Model):
     player = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    score = models.SmallIntegerField()
+    score = models.IntegerField()
     user_answer1 = models.TextField(default="")
     user_answer2 = models.TextField(default="")
     user_answer3 = models.TextField(default="")
